requirements.txt
asgiref==3.11.0
Django==6.0.1
django-debug-toolbar==4.4.6
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.12
factory_boy==3.3.3
Faker==40.1.2
inflection==0.5.1
iniconfig==2.3.0
packaging==25.0
pillow==12.1.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
pytest-factoryboy==2.8.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
python-telegram-bot==21.10
python-dotenv==1.0.1
